ADD : `Noun`(아이템 유형) 종류 추가.

+++ inventory-dto/src/main/kotlin/kr/lul/inventory/dto/NounSimpleDto.kt
@@ -1,3 +1,3 @@
 package kr.lul.inventory.dto
 
-data class NounSimpleDto(val id: Long, val key: String, val label: String, val labelCode: String)
+data class NounSimpleDto(val id: Int, val key: String, val label: String, val labelCode: String)
